Mandatory complete

+++ philo/src/monitor.c
@@ -6,7 +6,7 @@
 /*   By: dridolfo <dridolfo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/12 13:44:36 by dridolfo          #+#    #+#             */
-/*   Updated: 2022/04/14 16:57:18 by dridolfo         ###   ########.fr       */
+/*   Updated: 2022/04/19 15:22:21 by dridolfo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void	*monitor(void *argv)
 {
 	t_philo			*philo;
 	struct timeval	now;
-	long long		ms;
+	long			ms;
 
 	philo = argv;
 	while (!philo->env->end)
@@ -41,9 +41,9 @@ void	*monitor(void *argv)
 		gettimeofday(&now, NULL);
 		ms = time_to_ms(now) - time_to_ms(philo->last_time_to_eat);
 		gettimeofday(&now, NULL);
-		if (ms >= philo->env->time_to_die && philo->env->end == 0)
+		if (ms >= (long) philo->env->time_to_die && philo->env->end == 0)
 		{
-			printf("%lld\t%d\t %s\n", \
+			printf("%ld\t%d\t %s\n", \
 				time_to_ms(now) - time_to_ms(philo->env->born_time), \
 				philo->n + 1, "died");
 			philo->env->end = 1;
